feat(io): 新增BufferComparator解码器对比验证工具
新增功能：
- 实现双解码器输出对比验证（硬件 vs 软件）
- 支持YUV/RGB多格式自适应对比
- 2层验证策略：快速验证（PSNR-Y/G）→ 深度验证（PSNR-YUV/RGB）
- 感知加权：Y/G通道权重更高，符合人眼感知
- 详细验证报告：保存失败帧信息到文件

核心特性：
- 格式自适应：自动检测并选择最优对比算法
- 高性能：相同格式零转换，95%帧快速通过
- PSNR阈值：≥38dB通过，35-38dB警告，<35dB失败
- 支持本地文件和RTSP流对比测试

接口改进：
- WorkerBase::open()：新增无参数版本，从worker_config_读取参数
- IVideoFileNavigator::getBytesPerPixel()：返回值改为double，支持非整数字节（如NV12的1.5字节/像素）
- 移除已废弃的open(path, width, height, bpp)接口

+++ include/productionline/worker/BufferFillingWorkerFacade.hpp
@@ -182,8 +182,9 @@ public:
     
     /**
      * 获取每像素字节数
+     * @return 每像素字节数（浮点数，支持如NV12的1.5字节/像素）
      */
-    int getBytesPerPixel() const;
+    double getBytesPerPixel() const;
     
     /**
      * 获取文件路径
